tidy up the DataCreator code
