+++ DataCreator/DataCreator/AreaEntity.cs
@@ -5,15 +5,12 @@ namespace DataCreator
 {
     public record AreaEntity :SimpleArea
     {
-       
         public string AddressLine1 { get; set; }
         public string AddressLine2 { get; set; }
         public string AddressLine3 { get; set; }
         public string AddressLine4 { get; set; }
         public string Postcode { get; set; }
-
         public List<AreaRelation> ChildAreas { get; set; }= [];
-
     }
 
     public record SimpleArea
@@ -52,5 +49,4 @@ namespace DataCreator
         [JsonIgnore]
         public bool IsDirect { get; set; }
     }
-
  }
