+++ DataCreator/DataCreator/DataCreatorApplication.cs
@@ -6,22 +6,13 @@
 
         public async Task CreateDataAsync()
         {
-
             //we have zipped up the source data to save space to unzip to get them ready
+            DataFileReader.DeleteTempFiles();
 
-            try
-            {
-                DataFileManager.DeleteTempFiles();
-            }
-            catch
-            {
-                Console.WriteLine("Error deleting temporary files");
-            }
-
-            DataFileManager.UnzipSourceFiles();
+            DataFileReader.UnzipSourceFiles();
             Console.WriteLine("Unzipped source CSV files");
             //get the ids of the indicators chosen for PoC (about 30)
-            var pocIndicators=DataFileManager.GetPocIndicators();
+            var pocIndicators=DataFileReader.GetPocIndicators();
             Console.WriteLine($"Got {pocIndicators.Count} indicators");
             //get the age data - this is age range data
             var ageData= await _dataManager.GetAgeDataAsync();
@@ -31,13 +22,13 @@
             Console.WriteLine($"Created areas, we are using {areasWeWant.Count} areas");
             //create the health data
            
-            var areasAndIndicators=  DataManager.CreateHealthDataAndAgeData(areasWeWant, pocIndicators, ageData, yearFrom:2000, useIndicators:false);
+            var areasAndIndicators=  DataManager.CreateHealthDataAndAgeData(areasWeWant, pocIndicators, ageData);
             Console.WriteLine($"Created all health data");
             //create the indicator data
             await _dataManager.CreateIndicatorDataAsync(areasAndIndicators, pocIndicators, addAreasToIndicator: true);
             Console.WriteLine($"Created all health data");
             //clean up the unzipped files
-            DataFileManager.DeleteTempFiles();
+            DataFileReader.DeleteTempFiles();
             Console.WriteLine($"Deleted temp files");
         }
     }
